+++ lib/peddler/xml_response_parser.rb
@@ -3,6 +3,9 @@ require 'peddler/xml_parser'
 module Peddler
   # @api private
   class XMLResponseParser < XMLParser
+    MATCHER = /Message|Report|Result/
+    private_constant :MATCHER
+
     def next_token
       parse.fetch('NextToken', false)
     end
@@ -11,8 +14,7 @@ module Peddler
 
     def find_data
       payload = xml.values.first
-      found = payload.find { |k, _| k.include?('Result') } ||
-              payload.find { |k, _| k == 'Message' }
+      found = payload.find { |k, _| k.match(MATCHER) }
 
       found.last if found
     end